Split source control dropdown hit testing
