+++ src/app_core/native_shell/composition/state/hit_testing/chrome/source_controls.rs
@@ -3,6 +3,11 @@ use crate::app_core::native_shell::runtime_contract::{
     BrowserPillModel, BrowserPillState, FolderPaneIdModel,
 };
 
+#[path = "source_controls/dropdown.rs"]
+mod dropdown;
+
+pub(in crate::app_core::native_shell::composition::state) use dropdown::sidebar_filter_dropdown_spec;
+
 impl NativeShellState {
     /// Return the left-sidebar tag-editor input rect.
     pub(crate) fn sidebar_pill_editor_input_rect(
@@ -498,318 +503,6 @@ fn sidebar_filter_dropdown_facet_at_point(
         Some(SidebarFilterDropdownFacet::Key)
     } else {
         None
-    }
-}
-
-/// Build sidebar filter dropdown panel geometry and option buttons.
-pub(in crate::app_core::native_shell::composition::state) fn sidebar_filter_dropdown_spec(
-    layout: &ShellLayout,
-    style: &StyleTokens,
-    model: &AppModel,
-    dropdown: Option<SidebarFilterDropdownState>,
-) -> Option<(Rect, Vec<ActionButton>)> {
-    let dropdown = dropdown?;
-    let filters_rect = sidebar_workspace_sections(layout, style).filters;
-    let rows = sidebar_filter_row_rects(filters_rect, style.sizing);
-    let row = rows
-        .get(sidebar_filter_dropdown_row_index(dropdown.facet))
-        .copied()?;
-    let definitions = sidebar_filter_dropdown_buttons(dropdown.facet, model, style);
-    if definitions.is_empty() {
-        return None;
-    }
-    let sizing = style.sizing;
-    let panel_padding = sizing.panel_inset.max(4.0);
-    let button_width = row.width().max(132.0);
-    let button_height = sizing.sidebar_action_button_height.max(18.0);
-    let button_gap = sizing.sidebar_action_button_gap.max(2.0);
-    let panel_width = button_width + panel_padding * 2.0;
-    let panel_height = (button_height * definitions.len() as f32)
-        + (button_gap * definitions.len().saturating_sub(1) as f32)
-        + panel_padding * 2.0;
-    let min_x = layout.sidebar.min.x + sizing.panel_inset;
-    let max_x = (layout.sidebar.max.x - sizing.panel_inset - panel_width).max(min_x);
-    let below_y = row.max.y + sizing.border_width.max(1.0);
-    let above_y = row.min.y - panel_height - sizing.border_width.max(1.0);
-    let min_y = layout.sidebar.min.y + sizing.panel_inset;
-    let max_y = (layout.sidebar.max.y - sizing.panel_inset - panel_height).max(min_y);
-    let preferred_y = if below_y <= max_y { below_y } else { above_y };
-    let panel_min = Point::new(
-        row.min.x.clamp(min_x, max_x),
-        preferred_y.clamp(min_y, max_y),
-    );
-    let panel_rect = Rect::from_min_max(
-        panel_min,
-        Point::new(panel_min.x + panel_width, panel_min.y + panel_height),
-    );
-    let button_x = panel_rect.min.x + panel_padding;
-    let mut button_y = panel_rect.min.y + panel_padding;
-    let mut buttons = Vec::with_capacity(definitions.len());
-    for mut button in definitions {
-        button.rect = Rect::from_min_max(
-            Point::new(button_x, button_y),
-            Point::new(button_x + button_width, button_y + button_height),
-        );
-        buttons.push(button);
-        button_y += button_height + button_gap;
-    }
-    Some((panel_rect, buttons))
-}
-
-/// Return the rendered filter-row index for a dropdown facet.
-fn sidebar_filter_dropdown_row_index(facet: SidebarFilterDropdownFacet) -> usize {
-    match facet {
-        SidebarFilterDropdownFacet::Format => 0,
-        SidebarFilterDropdownFacet::BitDepth => 1,
-        SidebarFilterDropdownFacet::Channels => 2,
-        SidebarFilterDropdownFacet::Bpm => 3,
-        SidebarFilterDropdownFacet::Key => 4,
-        SidebarFilterDropdownFacet::Rating => 5,
-    }
-}
-
-/// Build the selectable option buttons for one sidebar filter dropdown.
-fn sidebar_filter_dropdown_buttons(
-    facet: SidebarFilterDropdownFacet,
-    model: &AppModel,
-    style: &StyleTokens,
-) -> Vec<ActionButton> {
-    use crate::app_core::app_api::state::{
-        BrowserBitDepthFacet, BrowserBpmFacet, BrowserChannelFacet, BrowserFormatFacet,
-        BrowserKeyFacet, BrowserSidebarFilterFacet, BrowserSidebarFilterOption,
-    };
-    match facet {
-        SidebarFilterDropdownFacet::Format => vec![
-            sidebar_toggle_button(
-                "WAV",
-                model
-                    .sidebar_filters
-                    .formats
-                    .contains(&BrowserFormatFacet::Wav),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Format(BrowserFormatFacet::Wav),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_clear_button(
-                BrowserSidebarFilterFacet::Format,
-                !model.sidebar_filters.formats.is_empty(),
-                style,
-            ),
-        ],
-        SidebarFilterDropdownFacet::BitDepth => vec![
-            sidebar_toggle_button(
-                "Unavailable",
-                model
-                    .sidebar_filters
-                    .bit_depths
-                    .contains(&BrowserBitDepthFacet::Unavailable),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::BitDepth(BrowserBitDepthFacet::Unavailable),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_clear_button(
-                BrowserSidebarFilterFacet::BitDepth,
-                !model.sidebar_filters.bit_depths.is_empty(),
-                style,
-            ),
-        ],
-        SidebarFilterDropdownFacet::Channels => vec![
-            sidebar_toggle_button(
-                "Mono",
-                model
-                    .sidebar_filters
-                    .channels
-                    .contains(&BrowserChannelFacet::Mono),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Channels(BrowserChannelFacet::Mono),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_toggle_button(
-                "Stereo",
-                model
-                    .sidebar_filters
-                    .channels
-                    .contains(&BrowserChannelFacet::Stereo),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Channels(BrowserChannelFacet::Stereo),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_toggle_button(
-                "Multi",
-                model
-                    .sidebar_filters
-                    .channels
-                    .contains(&BrowserChannelFacet::Multi),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Channels(BrowserChannelFacet::Multi),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_toggle_button(
-                "Unavailable",
-                model
-                    .sidebar_filters
-                    .channels
-                    .contains(&BrowserChannelFacet::Unavailable),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Channels(BrowserChannelFacet::Unavailable),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_clear_button(
-                BrowserSidebarFilterFacet::Channels,
-                !model.sidebar_filters.channels.is_empty(),
-                style,
-            ),
-        ],
-        SidebarFilterDropdownFacet::Bpm => vec![
-            sidebar_toggle_button(
-                "Unknown",
-                model
-                    .sidebar_filters
-                    .bpms
-                    .contains(&BrowserBpmFacet::Unknown),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Bpm(BrowserBpmFacet::Unknown),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_toggle_button(
-                "Slow <90",
-                model.sidebar_filters.bpms.contains(&BrowserBpmFacet::Slow),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Bpm(BrowserBpmFacet::Slow),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_toggle_button(
-                "Mid 90-129",
-                model.sidebar_filters.bpms.contains(&BrowserBpmFacet::Mid),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Bpm(BrowserBpmFacet::Mid),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_toggle_button(
-                "Fast 130+",
-                model.sidebar_filters.bpms.contains(&BrowserBpmFacet::Fast),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Bpm(BrowserBpmFacet::Fast),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_clear_button(
-                BrowserSidebarFilterFacet::Bpm,
-                !model.sidebar_filters.bpms.is_empty(),
-                style,
-            ),
-        ],
-        SidebarFilterDropdownFacet::Key => vec![
-            sidebar_toggle_button(
-                "Unknown",
-                model
-                    .sidebar_filters
-                    .keys
-                    .contains(&BrowserKeyFacet::Unknown),
-                UiAction::ToggleBrowserSidebarFilter {
-                    option: BrowserSidebarFilterOption::Key(BrowserKeyFacet::Unknown),
-                    additive: true,
-                },
-                style,
-            ),
-            sidebar_clear_button(
-                BrowserSidebarFilterFacet::Key,
-                !model.sidebar_filters.keys.is_empty(),
-                style,
-            ),
-        ],
-        SidebarFilterDropdownFacet::Rating => [-3, -2, -1, 0, 1, 2, 3, 4]
-            .into_iter()
-            .enumerate()
-            .map(|(index, level)| {
-                sidebar_toggle_button(
-                    rating_dropdown_label(level),
-                    model.browser.active_rating_filters[index],
-                    UiAction::ToggleBrowserRatingFilter {
-                        level,
-                        invert: false,
-                    },
-                    style,
-                )
-            })
-            .collect(),
-    }
-}
-
-/// Build a dropdown toggle button with active-state styling metadata.
-fn sidebar_toggle_button(
-    label: &'static str,
-    active: bool,
-    action: UiAction,
-    style: &StyleTokens,
-) -> ActionButton {
-    ActionButton {
-        rect: Rect::default(),
-        label,
-        icon: None,
-        enabled: true,
-        active,
-        action,
-        text_color: if active {
-            style.accent_mint
-        } else {
-            style.text_primary
-        },
-    }
-}
-
-/// Build a dropdown clear button for one sidebar filter facet.
-fn sidebar_clear_button(
-    facet: crate::app_core::app_api::state::BrowserSidebarFilterFacet,
-    enabled: bool,
-    style: &StyleTokens,
-) -> ActionButton {
-    ActionButton {
-        rect: Rect::default(),
-        label: "Clear",
-        icon: None,
-        enabled,
-        active: false,
-        action: UiAction::ClearBrowserSidebarFilter { facet },
-        text_color: if enabled {
-            style.accent_copper
-        } else {
-            style.text_muted
-        },
-    }
-}
-
-/// Return the user-facing rating dropdown label for one rating level.
-fn rating_dropdown_label(level: i8) -> &'static str {
-    match level {
-        -3 => "-3",
-        -2 => "-2",
-        -1 => "-1",
-        0 => "0",
-        1 => "1",
-        2 => "2",
-        3 => "3",
-        4 => "Locked",
-        _ => "",
     }
 }
